+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added [`moment-timezone`](https://www.npmjs.com/package/moment-timezone) package to JavaScript worker
 - Bump superblocksteam/run dep to v0.0.6
 - Include signing algorithms with verification keys when registering agent
+- Include signing algorithm and public key in response from Sign endpoint (`/v1/signature/sign`)
 
 ## v1.15.1